refactor(inference): centralize test helpers and align metrics verb naming
Review follow-ups on PR #159: move duplicated _sample/_hist/_make_png_bytes
test helpers into the shared _helpers.py module, and rename
track_request_start/track_request_end to observe_request_start/
observe_request_end on Metrics for consistency with the observe_* naming
used by every other recording method.
