+++ radiologist-inference/radiologist_inference_tests/test_input_and_error_metrics.py
@@ -30,37 +30,13 @@ Assertions compare scrape deltas (scrape -> act -> scrape -> compare), never
 absolute values.
 """
 
-import io
-from typing import Any, Tuple
+from typing import Any
 
-import numpy as np
 import pytest
-from _helpers import build_det_onnx, build_mcd_onnx
+from _helpers import _hist, _make_png_bytes, _sample, build_det_onnx, build_mcd_onnx
 from fastapi.testclient import TestClient
-from PIL import Image as PILImage
 
 from radiologist.inference import Classifier, Explainer, MCDropoutPredictor, create_app
-
-
-def _make_png_bytes(width: int = 64, height: int = 64) -> bytes:
-    arr = np.zeros((height, width, 3), dtype=np.uint8)
-    img = PILImage.fromarray(arr, mode="RGB")
-    buf = io.BytesIO()
-    img.save(buf, format="PNG")
-    return buf.getvalue()
-
-
-def _sample(body: str, name: str, **labels: str) -> float:
-    """Value of one sample line in a Prometheus exposition body, or 0.0."""
-    if labels:
-        rendered = ",".join(f'{k}="{v}"' for k, v in sorted(labels.items()))
-        prefix = f"{name}{{{rendered}}} "
-    else:
-        prefix = f"{name} "
-    for line in body.splitlines():
-        if line.startswith(prefix):
-            return float(line[len(prefix) :])
-    return 0.0
 
 
 def _errors(client: TestClient, route: str, error_type: str) -> float:
@@ -70,12 +46,6 @@ def _errors(client: TestClient, route: str, error_type: str) -> float:
         route=route,
         error_type=error_type,
     )
-
-
-def _hist(client: TestClient, name: str) -> Tuple[float, float]:
-    """Return (count, sum) for an unlabelled histogram at the current scrape."""
-    body = client.get("/metrics").text
-    return _sample(body, f"{name}_count"), _sample(body, f"{name}_sum")
 
 
 @pytest.fixture()
